ExJs: l'albero e l'interprete — il motore esegue
Due pezzi in un commit, perche' il secondo non si puo' provare senza il primo:
il costruttore dell'albero e l'interprete che ci cammina sopra. 122 prove, 0
sbagliate, e `var s=0;for(var i=1;i<=10;i++)s=s+i;s` fa 55.

-------------------------------------------------------------------------------
1. UNA FUNZIONE SOLA, E UNA TABELLA

Gli operatori binari di JavaScript sono una ventina su undici livelli di
precedenza. Scriverne una funzione per livello — `somma()` che chiama
`prodotto()` che chiama `unario()` — vuol dire undici funzioni quasi identiche,
e undici posti in cui sbagliare quando se ne aggiunge uno. Qui c'e' UNA
funzione che sale di livello guardando una tabella, e la tabella e' l'unico
posto dove la precedenza e' scritta.

-------------------------------------------------------------------------------
2. LE PROVE CONFRONTANO L'ALBERO STAMPATO, E NON E' PIGNOLERIA

La precedenza e' la cosa che si sbaglia IN SILENZIO. `1-2-3` con
l'associativita' invertita non da' nessun errore: da' 2 invece di -4, e il
numero diverso si vede tre schermate piu' in la'. L'unico modo di vedere un
albero storto e' guardarlo.

    1 - 2 - 3   ->  (binario - (binario - (numero 1) (numero 2)) (numero 3))
    a = b = 1   ->  (assegna = (nome a) (assegna = (nome b) (numero 1)))
    new a.b()   ->  (nuovo (membro (nome a) .b) (arg))

! IL TERZO E' IL PUNTO IN CUI I MOTORI SCRITTI IN FRETTA SBAGLIANO: `new`
lega piu' stretto della chiamata, quindi `new a.b()` costruisce `a.b`, non `a`.
Anche li' nessun errore, solo l'oggetto sbagliato. Percio' `new` legge la sua
coda SENZA le chiamate, prende gli argomenti, e solo dopo la coda ricomincia.

-------------------------------------------------------------------------------
3. IL PUNTO E VIRGOLA CHE NON C'E' SE LO METTE IL LINGUAGGIO

Non e' una comodita' da tollerare: e' una regola che CAMBIA IL SIGNIFICATO.

    function f(){return
    1;}          ->  (blocco (ritorna) (espr (numero 1)))

`return` seguito da un a capo rende `undefined` qualunque cosa ci sia sotto, e
la riga sotto diventa un'istruzione a se'. Un motore che la ignorasse
eseguirebbe programmi diversi da quelli scritti, e senza dire niente. Stessa
regola per il postfisso: `a` e `++b` su due righe sono DUE istruzioni.

-------------------------------------------------------------------------------
4. for..in E for CLASSICO COMINCIANO UGUALI

`for (var k in o)` e `for (var i = 0; ...)` non si distinguono finche' non si
e' letta l'inizializzazione. Percio' la si legge con `in` SPENTO come
operatore — e' il parametro `senza_in` che attraversa tutte le funzioni delle
espressioni — e poi si guarda se il gettone dopo e' `in`. Senza quello,
`for (var k in o)` verrebbe letto come un confronto.

-------------------------------------------------------------------------------
5. QUELLO CHE NON C'E' RIFIUTA, NON PRODUCE UN ALBERO STORTO

`switch`, `try/catch`, `throw`, le etichette, `with`, le espressioni regolari.
I gettoni ci sono gia' — li legge lex.c — ma qui danno un errore che DICE che
non ci sono:

    switch(a){}  ->  rifiutato: questo scaglione di ExJs non ha ancora
                     switch, try e throw

Un motore che analizza a meta' e' peggio di uno che rifiuta.

-------------------------------------------------------------------------------
6. 61 PROVE, 0 SBAGLIATE

    make prova-exjs      il banco, sull'host
    make verifica-exjs   e che compili anche per i386

! I DUE FALLIMENTI TROVATI PROVANDO ERANO NEL BANCO, NON NEL PARSER: `for..in`
usa tre figli su quattro e la stampa metteva un "-" nel buco; `return` senza
valore veniva atteso come `(ritorna -)` invece di `(ritorna)`. E' esattamente
il genere di cosa per cui vale la pena stampare l'albero invece di fidarsi.

-------------------------------------------------------------------------------
7. L'INTERPRETE (val.c, run.c)

! UN VALORE E' SESSANTAQUATTRO BIT, E DENTRO C'E' UN double VERO. La prima
stesura diceva «l'indice di una casella», e sarebbe stato piu' semplice — ma un
interprete che cammina sull'albero produce un valore per ogni operazione, e con
una casella per valore `for (i=0;i<10000;i++) x=x+1` chiederebbe ventimila
caselle per un ciclo che non conserva niente. Senza raccoglitore — e in questo
scaglione non c'e' — sarebbe un motore che muore sul primo ciclo vero. Con il
NaN-boxing numeri e booleani non costano niente; solo stringhe e oggetti
occupano una casella.

! UN AMBITO E' UN OGGETTO, e il suo prototipo e' l'ambito che lo racchiude. Non
e' un trucco per risparmiare codice: cercare una variabile e cercare una
proprieta' sono la stessa operazione lungo la stessa catena. Tenerli separati
vorrebbe dire due meccanismi che devono restare d'accordo, e le chiusure sono
proprio il punto in cui i due si toccano.

! DUE GUARDIE, E SONO OBBLIGATORIE IN UN BROWSER. `while(true){}` non deve poter
fermare il sistema e `function f(){f()}` non deve poter mangiare la pila del C.
Si contano i passi e la profondita', e quando si sfora si smette DICENDOLO:

    ricorsione senza fine  ->  troppe chiamate annidate (ricorsione senza fine?)

Un motore senza queste due si puo' usare solo su codice di cui ci si fida, cioe'
non sul web.

! LE CONVERSIONI SONO LA META' DEL LINGUAGGIO, E SONO DOVE STANNO LE SORPRESE.
Non si imitano per fedelta' cieca: sono cio' che le pagine vere si aspettano, e
sbagliarle produce numeri diversi senza dare nessun errore.

    1+'2'      -> "12"        '5'*2      -> 10
    1=='1'     -> true        1==='1'    -> false
    null+1     -> 1           undefined+1-> NaN
    if('0')    -> il ramo vero (e questo lo sbagliano in tanti)
    1/0        -> Infinity    (non un errore: fermerebbe pagine che funzionano)
    2147483648|0 -> -2147483648   (i bit lavorano su interi a 32 bit con segno)

! LE CHIUSURE FUNZIONANO, ed e' la prova che l'ambito e' fatto bene:

    function conta(){var n=0;return function(){return ++n;};}
    var a=conta(),b=conta(); a();a();b();   ->  1

I due contatori sono indipendenti perche' la funzione si ricorda l'ambito in cui
e' NATA, non quello in cui e' chiamata.

! E `undefined` E' UN NOME, NON UNA PAROLA CHIAVE — con `NaN` e `Infinity`.
Sembra un dettaglio da grammatica e non lo e': `x === undefined` e' il modo in
cui mezzo web controlla se una cosa esiste, e senza quelle tre righe dava «nome
non definito» su codice perfettamente normale. E' l'unico difetto vero trovato
dalle prove d'esecuzione.

-------------------------------------------------------------------------------
8. QUELLO CHE NON C'E', E VA SAPUTO

! NIENTE RACCOGLITORE DI MEMORIA. Un oggetto o una stringa presi non tornano
piu' indietro. E' sopportabile oggi perche' il NaN-boxing toglie di mezzo il
caso che conta e perche' un contesto vive quanto una pagina; diventa
insopportabile appena qualcuno concatena stringhe dentro un ciclo. Il posto per
un mark-sweep e' gia' preparato: gli oggetti stanno in un vettore e le radici
sono tre.

! E `exjs_chiama` DA FUORI NON C'E' ANCORA, e lo dice invece di rendere
`undefined` in silenzio: servirebbe uno stato d'esecuzione, che oggi vive dentro
exjs_esegui. Lo portera' lo scaglione della coda dei lavori, che ha lo stesso
bisogno — ed e' quello che serve a `setTimeout`.

+++ lib/exjs/exjs.h
@@ -29,10 +29,11 @@
  * -----------------------------------------------------------------------------
  * ! I VALORI SONO OPACHI, e si toccano solo attraverso queste funzioni.
  *
- * Un ExJsVal e' un numero, non una struttura da guardare dentro. Oggi e'
- * l'indice di una casella; domani puo' essere un NaN-boxing a 64 bit o un
- * puntatore con i bit bassi marcati. Chi lo tratta come opaco non se ne
- * accorge; chi ci guarda dentro va riscritto.
+ * Un ExJsVal e' un numero, non una struttura da guardare dentro. Oggi e' un
+ * NaN-boxing a 64 bit (il perche' e' scritto sotto, accanto al typedef);
+ * domani puo' essere un puntatore con i bit bassi marcati, o quello che
+ * QuickJS usa. Chi lo tratta come opaco non se ne accorge; chi ci guarda
+ * dentro va riscritto.
  *
  * -----------------------------------------------------------------------------
  * ! I BUFFER SONO DI CHI CHIAMA, come in exhtml e in excss.
@@ -94,7 +95,27 @@ extern "C" {
 #define EXJS_OGGETTO      5
 #define EXJS_FUNZIONE     6
 
-typedef unsigned int ExJsVal;       /* opaco: vedi in cima */
+/* =============================================================================
+ * ! UN VALORE E' SESSANTAQUATTRO BIT, E DENTRO C'E' UN double VERO.
+ *
+ * La prima stesura diceva «l'indice di una casella», e sarebbe stato piu'
+ * semplice — ma un interprete che cammina sull'albero produce un valore per
+ * ogni operazione, e con una casella per valore `for (i=0;i<10000;i++) x=x+1`
+ * chiederebbe ventimila caselle per un ciclo che non conserva niente. Senza un
+ * raccoglitore di memoria — e in questo scaglione non c'e' — sarebbe un motore
+ * che muore sul primo ciclo vero.
+ *
+ * Percio' NaN-boxing: un valore E' un double, e i valori che double non sono
+ * si nascondono dentro i NaN silenziosi, che nella norma IEEE 754 sono
+ * duemiladuecentocinquantatre miliardi di configurazioni tutte uguali fra
+ * loro. Numeri e booleani non costano niente; solo stringhe e oggetti
+ * occupano una casella.
+ *
+ * ! RESTA OPACO. Chi lo tratta come un numero da confrontare con exjs_tipo()
+ * non si accorgera' del giorno che ci mettiamo QuickJS sotto; chi ci guarda
+ * dentro va riscritto. Le macro che lo aprono stanno in exjs_int.h, non qui.
+ * ========================================================================== */
+typedef unsigned long long ExJsVal;
 
 /* Il contesto: tutto lo stato di un'esecuzione. Chi chiama lo alloca e ne
  * decide la taglia; la libreria non ne tiene nessuno per conto suo. */
